+++ packages/expo-modules-core/android/src/main/cpp/JSIInteropModuleRegistry.h
@@ -64,12 +64,16 @@ public:
    */
   jni::local_ref<JavaScriptObject::javaobject> global();
 
+  /**
+   * Exposes a `JavaScriptRuntime::createObject` function to Kotlin
+   */
+  jni::local_ref<JavaScriptObject::javaobject> createObject();
+
   std::shared_ptr<react::CallInvoker> jsInvoker;
   std::shared_ptr<react::CallInvoker> nativeInvoker;
-
+  std::shared_ptr<JavaScriptRuntime> runtimeHolder;
 private:
   friend HybridBase;
-  std::shared_ptr<JavaScriptRuntime> runtimeHolder;
   jni::global_ref<JSIInteropModuleRegistry::javaobject> javaPart_;
 
   explicit JSIInteropModuleRegistry(jni::alias_ref<jhybridobject> jThis);