+++ packages/expo-modules-core/android/src/main/cpp/JavaScriptObject.h
@@ -2,6 +2,10 @@
 
 #pragma once
 
+#include "JSIObjectWrapper.h"
+#include "JSITypeConverter.h"
+#include "JavaScriptRuntime.h"
+
 #include <fbjni/fbjni.h>
 #include <jsi/jsi.h>
 
@@ -18,7 +22,7 @@ class JavaScriptRuntime;
 /**
  * Represents any JavaScript object. Its purpose is to exposes `jsi::Object` API back to Kotlin.
  */
-class JavaScriptObject : public jni::HybridClass<JavaScriptObject> {
+class JavaScriptObject : public jni::HybridClass<JavaScriptObject>, JSIObjectWrapper {
 public:
   static auto constexpr
     kJavaDescriptor = "Lexpo/modules/kotlin/jni/JavaScriptObject;";
@@ -30,6 +34,8 @@ public:
     std::weak_ptr<JavaScriptRuntime> runtime,
     std::shared_ptr<jsi::Object> jsObject
   );
+
+  std::shared_ptr<jsi::Object> get() override;
 
   /**
    * @return a bool whether the object has a property with the given name
@@ -47,6 +53,8 @@ public:
    */
   std::vector<std::string> getPropertyNames();
 
+  void setProperty(const std::string &name, jsi::Value value);
+
 private:
   friend HybridBase;
   std::weak_ptr<JavaScriptRuntime> runtimeHolder;
@@ -59,5 +67,65 @@ private:
   );
 
   jni::local_ref<jni::JArrayClass<jstring>> jniGetPropertyNames();
+
+  /**
+   * Unsets property with the given name.
+   */
+  void unsetProperty(jni::alias_ref<jstring> name);
+
+  /**
+   * A template to generate different versions of the `setProperty` method based on the `jsi_type_converter` trait.
+   * Those generated methods will be exported and visible in the Kotlin codebase.
+   * On the other hand, we could just make one function that would take a generic Java Object,
+   * but then we would have to decide what to do with it and how to convert it to jsi::Value
+   * in cpp. That would be expensive. So it's easier to ensure that
+   * we call the correct version of `setProperty` in the Kotlin code.
+   *
+   * This template will work only if the jsi_type_converter exists for a given type.
+   */
+  template<
+    class T,
+    typename = std::enable_if_t<is_jsi_type_converter_defined<T>>
+  >
+  void setProperty(jni::alias_ref<jstring> name, T value) {
+    auto runtime = runtimeHolder.lock();
+    assert(runtime != nullptr);
+    auto cName = name->toStdString();
+
+    jsObject->setProperty(
+      *runtime->get(),
+      cName.c_str(),
+      jsi_type_converter<T>::convert(*runtime->get(), value)
+    );
+  }
+
+  template<
+    class T,
+    typename = std::enable_if_t<is_jsi_type_converter_defined<T>>
+  >
+  void defineProperty(jni::alias_ref<jstring> name, T value, int options) {
+    auto runtime = runtimeHolder.lock();
+    assert(runtime != nullptr);
+    jsi::Runtime &jsRuntime = *runtime->get();
+
+    auto cName = name->toStdString();
+    jsi::Object global = jsRuntime.global();
+    jsi::Object objectClass = global.getPropertyAsObject(jsRuntime, "Object");
+    jsi::Function definePropertyFunction = objectClass.getPropertyAsFunction(
+      jsRuntime,
+      "defineProperty"
+    );
+    jsi::Object descriptor = preparePropertyDescriptor(jsRuntime, options);
+
+    descriptor.setProperty(jsRuntime, "value", jsi_type_converter<T>::convert(jsRuntime, value));
+
+    definePropertyFunction.callWithThis(jsRuntime, objectClass, {
+      jsi::Value(jsRuntime, *jsObject),
+      jsi::String::createFromUtf8(jsRuntime, cName),
+      std::move(descriptor)
+    });
+  }
+
+  static jsi::Object preparePropertyDescriptor(jsi::Runtime &jsRuntime, int options);
 };
 } // namespace expo