+++ packages/expo-modules-core/android/src/main/cpp/JavaScriptValue.h
@@ -2,7 +2,7 @@
 
 #pragma once
 
-#include "JavaScriptObject.h"
+#include "JSIObjectWrapper.h"
 
 #include <fbjni/fbjni.h>
 #include <jsi/jsi.h>
@@ -15,10 +15,12 @@ namespace jsi = facebook::jsi;
 namespace expo {
 class JavaScriptRuntime;
 
+class JavaScriptObject;
+
 /**
  * Represents any JavaScript value. Its purpose is to expose the `jsi::Value` API back to Kotlin.
  */
-class JavaScriptValue : public jni::HybridClass<JavaScriptValue> {
+class JavaScriptValue : public jni::HybridClass<JavaScriptValue>, JSIValueWrapper {
 public:
   static auto constexpr
     kJavaDescriptor = "Lexpo/modules/kotlin/jni/JavaScriptValue;";
@@ -31,28 +33,42 @@ public:
     std::shared_ptr<jsi::Value> jsValue
   );
 
+  std::shared_ptr<jsi::Value> get() override;
+
   std::string kind();
 
   bool isNull();
+
   bool isUndefined();
+
   bool isBool();
+
   bool isNumber();
+
   bool isString();
+
   bool isSymbol();
+
   bool isFunction();
+
   bool isObject();
 
   bool getBool();
+
   double getDouble();
+
   std::string getString();
-  jni::local_ref<JavaScriptObject::javaobject> getObject();
+
+  jni::local_ref<jni::HybridClass<JavaScriptObject>::javaobject> getObject();
 
 private:
   friend HybridBase;
+
   std::weak_ptr<JavaScriptRuntime> runtimeHolder;
   std::shared_ptr<jsi::Value> jsValue;
 
   jni::local_ref<jstring> jniKind();
+
   jni::local_ref<jstring> jniGetString();
 };
 } // namespace expo